# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  /api/auth/register:
    post:
      summary: ''
      operationId: postApiAuthRegister
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: hvswyunuqfqndenzmwnhpi
                  nullable: false
                phone:
                  type: number
                  description: ''
                  example: 3.3
                  nullable: false
              required:
                - full_name
                - phone
      security: []
  /api/auth/send-otp:
    post:
      summary: ''
      operationId: postApiAuthSendOtp
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  description: 'The <code>phone</code> of an existing record in the users table.'
                  example: culpa
                  nullable: false
              required:
                - phone
      security: []
  /api/auth/verify-otp:
    post:
      summary: ''
      operationId: postApiAuthVerifyOtp
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  description: 'The <code>phone</code> of an existing record in the users table.'
                  example: dolorum
                  nullable: false
                otp:
                  type: number
                  description: ''
                  example: 2.366437
                  nullable: false
              required:
                - phone
                - otp
      security: []
...
  /api/auth/profile-image:
    post:
      summary: ''
      operationId: postApiAuthProfileImage
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: 'Must be an image. Must not be greater than 2024 kilobytes.'
                  nullable: false
              required:
                - image
      security: []
...
  '/api/room/{room_id}/c4/start':
    post:
      summary: ''
      operationId: postApiRoomRoom_idC4Start
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: room_id
        description: 'The ID of the room.'
        example: 1
        required: true
        schema:
          type: integer
  '/api/room/{room_id}/c4/{gameId}/make-move':
    post:
      summary: ''
      operationId: postApiRoomRoom_idC4GameIdMakeMove
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                column:
                  type: integer
                  description: 'Must be at least 0. Must not be greater than 6.'
                  example: 3
                  nullable: false
              required:
                - column
      security: []
    parameters:
      -
        in: path
        name: room_id
        description: 'The ID of the room.'
        example: 1
        required: true
        schema:
          type: integer
      -
        in: path
        name: gameId
        description: ''
        example: est
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []